GodSight/api/requirements.txt
Flask
psycopg2-binary
psycopg2
python-dotenv
pandas
flask_sqlalchemy
requests
flasgger